+++ Changelog.md
@@ -2,6 +2,8 @@
 
 ## 0.6.6 ()
 
+- `HierarchicalMachine` now considers the initial state of `NestedState` instances/names passed to `initial`.
+
 ## 0.6.5 (April, 2018)
 
 Release 0.6.5 is a minor release and contains a new feature and a bugfix: